move csgraph tests into subdirectory
